requirements.txt
ultralytics>=8.0.0
opencv-python>=4.7.0
numpy>=1.22.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
pillow>=9.4.0
mediapipe>=0.10.0
requests>=2.28.0
pyyaml>=6.0
